removed the global variables :rage1:
Making global variables the way I was doing before could lead to problems later. For best practice it's better to encapsulate them in a namespace and also using "const" to make sure the variables have internal linkage. I also passed the constants as parameters to other functions to improve modularity of the code (loose Coupling)