Linkedlist demo now searches with the triple pointer technique :godmode:
I wanted to try the method of using a pointer that points to a pointer of a list (triple pointer method) after I read a bit about it. After giving it some thought, I decided the best way of doing this was to implement a simple linear search on my linked list. Now you can ask the demo where on the list a certain element is and it will tell you where it is (or if it couldn't find it on the list) :godmode: :+1: